Revbump, fixing bug #296221.
Package-Manager: portage-2.2_rc78/cvs/Linux x86_64

+++ net-print/cups/cups-1.4.4-r2.ebuild
@@ -1,10 +1,13 @@
 # Copyright 1999-2010 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/net-print/cups/cups-1.4.4.ebuild,v 1.3 2010/07/30 22:32:11 anarchy Exp $
+# $Header: /var/cvsroot/gentoo-x86/net-print/cups/cups-1.4.4-r2.ebuild,v 1.1 2010/09/11 12:10:19 tgurr Exp $
 
-EAPI="2"
+EAPI=3
 
-inherit autotools eutils flag-o-matic multilib pam versionator
+PYTHON_DEPEND="python? 2:2.5"
+RESTRICT_PYTHON_ABIS="3.*"
+
+inherit autotools eutils flag-o-matic multilib pam perl-module python versionator java-pkg-opt-2
 
 MY_P=${P/_}
 
@@ -15,9 +18,14 @@ SRC_URI="mirror://easysw/${PN}/${PV}/${MY_P}-source.tar.bz2"
 LICENSE="GPL-2"
 SLOT="0"
 KEYWORDS="~alpha ~amd64 ~arm ~hppa ~ia64 ~m68k ~mips ~ppc ~ppc64 ~s390 ~sh ~sparc ~sparc-fbsd ~x86 ~x86-fbsd"
-IUSE="acl dbus debug gnutls java +jpeg kerberos ldap pam perl php +png python samba slp +ssl static +threads +tiff +usb X xinetd"
+IUSE="acl dbus debug gnutls java +jpeg kerberos ldap pam perl php +png python samba slp +ssl static-libs +threads +tiff +usb X xinetd"
 
-COMMON_DEPEND="
+LANGS="da de es eu fi fr id it ja ko nl no pl pt pt_BR ru sv zh zh_TW"
+for X in ${LANGS} ; do
+	IUSE="${IUSE} linguas_${X}"
+done
+
+RDEPEND="
 	app-text/libpaper
 	dev-libs/libgcrypt
 	acl? (
@@ -27,31 +35,30 @@ COMMON_DEPEND="
 		)
 	)
 	dbus? ( sys-apps/dbus )
-	gnutls? ( net-libs/gnutls )
-	java? ( >=virtual/jre-1.4 )
-	jpeg? ( >=media-libs/jpeg-6b:0 )
+	java? ( >=virtual/jre-1.6 )
+	jpeg? ( virtual/jpeg:0 )
 	kerberos? ( virtual/krb5 )
-	ldap? ( net-nds/openldap )
+	ldap? ( net-nds/openldap[ssl?,gnutls?] )
 	pam? ( virtual/pam )
 	perl? ( dev-lang/perl )
 	php? ( dev-lang/php )
-	png? ( >=media-libs/libpng-1.2.1 )
-	python? ( dev-lang/python )
+	png? ( >=media-libs/libpng-1.4.3 )
 	slp? ( >=net-libs/openslp-1.0.4 )
 	ssl? (
+		gnutls? ( net-libs/gnutls )
 		!gnutls? ( >=dev-libs/openssl-0.9.8g )
 	)
 	tiff? ( >=media-libs/tiff-3.5.5 )
-	usb? ( dev-libs/libusb )
-	xinetd? ( sys-apps/xinetd )
-"
-DEPEND="${COMMON_DEPEND}"
-
-RDEPEND="${COMMON_DEPEND}
-	!net-print/cupsddk
-	!virtual/lpr
+	usb? ( virtual/libusb:0 )
 	X? ( x11-misc/xdg-utils )
+	xinetd? ( sys-apps/xinetd )
+	!net-print/cupsddk
 "
+
+DEPEND="${RDEPEND}
+	dev-util/pkgconfig
+"
+
 PDEPEND="
 	app-text/ghostscript-gpl[cups]
 	>=app-text/poppler-0.12.3-r3[utils]
@@ -60,31 +67,28 @@ PDEPEND="
 PROVIDE="virtual/lpr"
 
 # upstream includes an interactive test which is a nono for gentoo.
-# therefore, since the printing herd has bigger fish to fry, for now,
-# we just leave it out, even if FEATURES=test
 RESTRICT="test"
 
 S="${WORKDIR}/${MY_P}"
-
-LANGS="da de es eu fi fr id it ja ko nl no pl pt pt_BR ru sv zh zh_TW"
-for X in ${LANGS} ; do
-	IUSE="${IUSE} linguas_${X}"
-done
 
 pkg_setup() {
 	enewgroup lp
 	enewuser lp -1 -1 -1 lp
 	enewgroup lpadmin 106
+
+	# python 3 is no-go
+	python_set_active_version 2
+	python_pkg_setup
 }
 
 src_prepare() {
-	# remove default optimizations and do not strip by default
-	sed -e 's:OPTIM="-Os -g":OPTIM="":' \
-		-e 's:INSTALL_STRIP="-s":INSTALL_STRIP="":' \
-		-i config-scripts/cups-compiler.m4
-
 	# create a missing symlink to allow https printing via IPP, bug #217293
 	epatch "${FILESDIR}/${PN}-1.4.0-backend-https.patch"
+	# various build time fixes
+	epatch "${FILESDIR}/${PN}-1.4.4-dont-compress-manpages.patch"
+	epatch "${FILESDIR}/${PN}-1.4.4-fix-install-perms.patch"
+	epatch "${FILESDIR}/${PN}-1.4.4-nostrip.patch"
+	epatch "${FILESDIR}/${PN}-1.4.4-php-destdir.patch"
 
 	AT_M4DIR=config-scripts eaclocal
 	eautoconf
@@ -101,13 +105,15 @@ src_configure() {
 
 	local myconf
 	if use ssl || use gnutls ; then
-		myconf="${myconf} \
-			$(use_enable gnutls) \
-			$(use_enable !gnutls openssl)"
+		myconf+="
+			$(use_enable gnutls)
+			$(use_enable !gnutls openssl)
+		"
 	else
-		myconf="${myconf} \
-			--disable-gnutls \
-			--disable-openssl"
+		myconf+="
+			--disable-gnutls
+			--disable-openssl
+		"
 	fi
 
 	econf \
@@ -129,7 +135,7 @@ src_configure() {
 		$(use_enable pam) \
 		$(use_enable png) \
 		$(use_enable slp) \
-		$(use_enable static) \
+		$(use_enable static-libs static) \
 		$(use_enable threads) \
 		$(use_enable tiff) \
 		$(use_enable usb libusb) \
@@ -144,23 +150,52 @@ src_configure() {
 
 	# install in /usr/libexec always, instead of using /usr/lib/cups, as that
 	# makes more sense when facing multilib support.
-	sed -i -e 's:SERVERBIN.*:SERVERBIN = "$(BUILDROOT)"/usr/libexec/cups:' Makedefs
-	sed -i -e 's:#define CUPS_SERVERBIN.*:#define CUPS_SERVERBIN "/usr/libexec/cups":' config.h
-	sed -i -e 's:cups_serverbin=.*:cups_serverbin=/usr/libexec/cups:' cups-config
+	sed -i -e 's:SERVERBIN.*:SERVERBIN = "$(BUILDROOT)"/usr/libexec/cups:' Makedefs || die
+	sed -i -e 's:#define CUPS_SERVERBIN.*:#define CUPS_SERVERBIN "/usr/libexec/cups":' config.h || die
+	sed -i -e 's:cups_serverbin=.*:cups_serverbin=/usr/libexec/cups:' cups-config || die
+}
+
+src_compile() {
+	emake || die "emake failed"
+
+	if use perl ; then
+		cd "${S}"/scripting/perl
+		perl-module_src_prep
+		perl-module_src_compile
+	fi
+
+	if use php ; then
+		cd "${S}"/scripting/php
+		emake || die "emake php failed"
+	fi
 }
 
 src_install() {
 	emake BUILDROOT="${D}" install || die "emake install failed"
 	dodoc {CHANGES,CREDITS,README}.txt || die "dodoc install failed"
 
+	if use perl ; then
+		cd "${S}"/scripting/perl
+		perl-module_src_install
+		fixlocalpod
+	fi
+
+	if use php ; then
+		cd "${S}"/scripting/php
+		emake DESTDIR="${D}" install || die "emake install for php bindings failed"
+	fi
+
 	# clean out cups init scripts
 	rm -rf "${D}"/etc/{init.d/cups,rc*,pam.d/cups}
 
 	# install our init script
 	local neededservices
-	use dbus && neededservices="$neededservices dbus"
+	use dbus && neededservices+=" dbus"
 	[[ -n ${neededservices} ]] && neededservices="need${neededservices}"
-	sed -e "s/@neededservices@/$neededservices/" "${FILESDIR}"/cupsd.init.d > "${T}"/cupsd
+	cp "${FILESDIR}"/cupsd.init.d "${T}"/cupsd || die
+	sed -i \
+		-e "s/@neededservices@/$neededservices/" \
+		"${T}"/cupsd || die
 	doinitd "${T}"/cupsd || die "doinitd failed"
 
 	# install our pam script
@@ -168,10 +203,14 @@ src_install() {
 
 	if use xinetd ; then
 		# correct path
-		sed -i -e "s:server = .*:server = /usr/libexec/cups/daemon/cups-lpd:" "${D}"/etc/xinetd.d/cups-lpd
+		sed -i \
+			-e "s:server = .*:server = /usr/libexec/cups/daemon/cups-lpd:" \
+			"${D}"/etc/xinetd.d/cups-lpd || die
 		# it is safer to disable this by default, bug #137130
 		grep -w 'disable' "${D}"/etc/xinetd.d/cups-lpd || \
-			sed -i -e "s:}:\tdisable = yes\n}:" "${D}"/etc/xinetd.d/cups-lpd
+			{ sed -i -e "s:}:\tdisable = yes\n}:" "${D}"/etc/xinetd.d/cups-lpd || die ; }
+		# write permission for file owner (root), bug #296221
+		fperms u+w /etc/xinetd.d/cups-lpd || die "fperms failed"
 	else
 		rm -rf "${D}"/etc/xinetd.d
 	fi
